pyproject.toml
[tool.poetry]
name = "markdata"
version = "0.0.1"
description = "Tool to convert markdown tables into json objects"
authors = ["Gordon Bean <gbean@cs.byu.edu>", "Robert Greathouse <robbykap@byu.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
